+++ rust/dbn/src/encode/json.rs
@@ -126,7 +126,10 @@ mod tests {
     use crate::{
         encode::test_data::{VecStream, BID_ASK, RECORD_HEADER},
         enums::{SType, Schema, SecurityUpdateAction},
-        record::{InstrumentDefMsg, MboMsg, Mbp10Msg, Mbp1Msg, OhlcvMsg, StatusMsg, TradeMsg},
+        record::{
+            str_to_c_chars, InstrumentDefMsg, MboMsg, Mbp10Msg, Mbp1Msg, OhlcvMsg, StatusMsg,
+            TradeMsg,
+        },
         MappingInterval, SymbolMapping,
     };
 
@@ -168,7 +171,7 @@ mod tests {
     const BID_ASK_JSON: &str = r#"{"bid_px":372000000000000,"ask_px":372500000000000,"bid_sz":10,"ask_sz":5,"bid_ct":5,"ask_ct":2}"#;
 
     #[test]
-    fn test_tick_write_json() {
+    fn test_mbo_write_json() {
         let data = vec![MboMsg {
             hd: RECORD_HEADER,
             order_id: 16,
@@ -176,8 +179,8 @@ mod tests {
             size: 3,
             flags: 128,
             channel_id: 14,
-            action: 'B' as i8,
-            side: 67,
+            action: 'R' as c_char,
+            side: 'N' as c_char,
             ts_recv: 1658441891000000000,
             ts_in_delta: 22_000,
             sequence: 1_002_375,
@@ -190,19 +193,19 @@ mod tests {
             slice_res,
             format!(
                 "{{{HEADER_JSON},{}}}\n",
-                r#""order_id":16,"price":5500,"size":3,"flags":128,"channel_id":14,"action":66,"side":67,"ts_recv":"1658441891000000000","ts_in_delta":22000,"sequence":1002375"#
+                r#""order_id":16,"price":5500,"size":3,"flags":128,"channel_id":14,"action":"R","side":"N","ts_recv":"1658441891000000000","ts_in_delta":22000,"sequence":1002375"#
             )
         );
     }
 
     #[test]
-    fn test_mbo1_write_json() {
+    fn test_mbp1_write_json() {
         let data = vec![Mbp1Msg {
             hd: RECORD_HEADER,
             price: 5500,
             size: 3,
-            action: 'B' as i8,
-            side: 67,
+            action: 'B' as c_char,
+            side: 'B' as c_char,
             flags: 128,
             depth: 9,
             ts_recv: 1658441891000000000,
@@ -218,7 +221,7 @@ mod tests {
             slice_res,
             format!(
                 "{{{HEADER_JSON},{},{}}}\n",
-                r#""price":5500,"size":3,"action":66,"side":67,"flags":128,"depth":9,"ts_recv":"1658441891000000000","ts_in_delta":22000,"sequence":1002375"#,
+                r#""price":5500,"size":3,"action":"B","side":"B","flags":128,"depth":9,"ts_recv":"1658441891000000000","ts_in_delta":22000,"sequence":1002375"#,
                 format_args!("\"booklevel\":[{BID_ASK_JSON}]")
             )
         );
@@ -230,8 +233,8 @@ mod tests {
             hd: RECORD_HEADER,
             price: 5500,
             size: 3,
-            action: 'B' as i8,
-            side: 67,
+            action: 'T' as c_char,
+            side: 'N' as c_char,
             flags: 128,
             depth: 9,
             ts_recv: 1658441891000000000,
@@ -247,7 +250,7 @@ mod tests {
             slice_res,
             format!(
                 "{{{HEADER_JSON},{},{}}}\n",
-                r#""price":5500,"size":3,"action":66,"side":67,"flags":128,"depth":9,"ts_recv":"1658441891000000000","ts_in_delta":22000,"sequence":1002375"#,
+                r#""price":5500,"size":3,"action":"T","side":"N","flags":128,"depth":9,"ts_recv":"1658441891000000000","ts_in_delta":22000,"sequence":1002375"#,
                 format_args!("\"booklevel\":[{BID_ASK_JSON},{BID_ASK_JSON},{BID_ASK_JSON},{BID_ASK_JSON},{BID_ASK_JSON},{BID_ASK_JSON},{BID_ASK_JSON},{BID_ASK_JSON},{BID_ASK_JSON},{BID_ASK_JSON}]")
             )
         );
@@ -259,8 +262,8 @@ mod tests {
             hd: RECORD_HEADER,
             price: 5500,
             size: 3,
-            action: 'B' as i8,
-            side: 67,
+            action: 'C' as c_char,
+            side: 'B' as c_char,
             flags: 128,
             depth: 9,
             ts_recv: 1658441891000000000,
@@ -276,7 +279,7 @@ mod tests {
             slice_res,
             format!(
                 "{{{HEADER_JSON},{}}}\n",
-                r#""price":5500,"size":3,"action":66,"side":67,"flags":128,"depth":9,"ts_recv":"1658441891000000000","ts_in_delta":22000,"sequence":1002375"#,
+                r#""price":5500,"size":3,"action":"C","side":"B","flags":128,"depth":9,"ts_recv":"1658441891000000000","ts_in_delta":22000,"sequence":1002375"#,
             )
         );
     }
@@ -369,7 +372,7 @@ mod tests {
             decay_start_date: 0,
             channel_id: 4,
             currency: [0; 4],
-            settl_currency: ['U' as c_char, 'S' as c_char, 'D' as c_char, 0],
+            settl_currency: str_to_c_chars("USD").unwrap(),
             secsubtype: [0; 6],
             symbol: [0; 22],
             group: [0; 21],
@@ -380,7 +383,7 @@ mod tests {
             unit_of_measure: [0; 31],
             underlying: [0; 21],
             related: [0; 21],
-            match_algorithm: 1,
+            match_algorithm: 'F' as c_char,
             md_security_trading_status: 2,
             main_fraction: 4,
             price_display_format: 8,
@@ -391,7 +394,7 @@ mod tests {
             maturity_month: 8,
             maturity_day: 9,
             maturity_week: 11,
-            user_defined_instrument: 1,
+            user_defined_instrument: 'N' as c_char,
             contract_multiplier_unit: 0,
             flow_schedule_type: 5,
             tick_rule: 0,
@@ -412,9 +415,9 @@ mod tests {
                     r#""market_depth":13,"market_segment_id":0,"max_trade_vol":10000,"min_lot_size":1,"min_lot_size_block":1000,"min_lot_size_round_lot":100,"min_trade_vol":1,"#,
                     r#""open_interest_qty":0,"contract_multiplier":0,"decay_quantity":0,"original_contract_size":0,"related_security_id":0,"trading_reference_date":0,"appl_id":0,"#,
                     r#""maturity_year":0,"decay_start_date":0,"channel_id":4,"currency":"","settl_currency":"USD","secsubtype":"","symbol":"","group":"","exchange":"","asset":"","cfi":"","#,
-                    r#""security_type":"","unit_of_measure":"","underlying":"","related":"","match_algorithm":1,"md_security_trading_status":2,"main_fraction":4,"price_display_format":8,"#,
+                    r#""security_type":"","unit_of_measure":"","underlying":"","related":"","match_algorithm":"F","md_security_trading_status":2,"main_fraction":4,"price_display_format":8,"#,
                     r#""settl_price_type":9,"sub_fraction":23,"underlying_product":10,"security_update_action":"A","maturity_month":8,"maturity_day":9,"maturity_week":11,"#,
-                    r#""user_defined_instrument":1,"contract_multiplier_unit":0,"flow_schedule_type":5,"tick_rule":0"#
+                    r#""user_defined_instrument":"N","contract_multiplier_unit":0,"flow_schedule_type":5,"tick_rule":0"#
                 )
             )
         );